stopjudol/requirements.txt
# Client Dependencies
PyQt6>=6.4.0
google-api-python-client>=2.70.0
google-auth-oauthlib>=1.0.0
google-auth>=2.16.0
keyring>=23.13.1
requests>=2.28.2
jsonrpcclient>=4.0.0

# Server Dependencies
jsonrpcserver>=5.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
marshmallow>=3.19.0
pyjwt>=2.6.0

# Development Dependencies
pyinstaller>=6.0.0
